Added loyalty repo & services

+++ src/main/java/j2ee_project/dao/UserRepository.java
@@ -5,7 +5,7 @@ import j2ee_project.model.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> getUserById(Long id);
+    Optional<User> getUserById(int id);
 //    @Query("select role from Role role")
 //    Stream<Role> getAllRolesStream();
 }
