+++ src/main/java/j2ee_project/dao/catalog/category/CategoryRepository.java
@@ -11,6 +11,6 @@ import java.util.Optional;
  * The interface Category repository to interacts with the Category table in the DB
  */
 public interface CategoryRepository extends JpaRepository<Category, Long> {
-    Category findCategoryBy(Long id);
-    void deleteCategoryById(Long id);
+    Category findCategoryBy(int id);
+    void deleteCategoryById(int id);
 }
