+++ src/main/java/j2ee_project/dao/discount/DiscountRepository.java
@@ -4,6 +4,6 @@ import j2ee_project.model.Discount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiscountRepository extends JpaRepository<Discount,Long> {
-    public Discount findDiscountById(Long discountId);
-    public void deleteDiscountById(Long discountId);
+    public Discount findDiscountById(int discountId);
+    public void deleteDiscountById(int discountId);
 }
